fix: dont reset token to default when click setting
